- switched on tags to data attributes to hold information about action and entity id - game views started - game overview ready (missing game date for now) - Added gameDate to Game entity - Added test data on created on startup - added game link to header - transaction advice added for repositories - view model created for game

+++ 2016-05-27_hands_on/parent/model/view/src/main/java/at/fh/ooe/swt6/em/model/view/team/TeamView.java
@@ -17,9 +17,9 @@ public class TeamView implements Serializable {
     //<editor-fold desc="Properties">
     Long id;
     String name;
-    Long gamesWin;
-    Long gamesLost;
-    Long gamesEqual;
+    Integer gamesWin;
+    Integer gamesLost;
+    Integer gamesEqual;
     //</editor-fold>
 
     public TeamView(Long id) {
